Refactor: update imports of `updateMotionValues`
