+++ src/modules/dashboard/sagas/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './updateColonyDisplayCache';
+export { updateMotionValues } from './updateMotionValues';
